Split up `EventStreamTestTools`
